+++ src/mfgparams/__init__.py
@@ -69,4 +69,4 @@ __all__ = [
     "MillingSubOperation",
 ]
 
-__version__ = "2.4.0"
+__version__ = "2.5.0"